Factor some functions out of nodes.
