+++ nodes/vector/attractor.py
@@ -22,29 +22,15 @@ from mathutils import Vector, Matrix
 from mathutils.kdtree import KDTree
 import math
 
-from sverchok.node_tree import SverchCustomTreeNode
+from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, match_long_repeat, fullList
+from sverchok.utils.math import inverse, inverse_square, inverse_cubic, inverse_exp, gauss
 
 def get_avg_vector(vectors):
     result = Vector((0,0,0))
     for vector in vectors:
         result += vector
     return result / len(vectors)
-
-def inverse(c, x):
-    return 1.0/x
-
-def inverse_square(c, x):
-    return 1.0/(x*x)
-
-def inverse_cubic(c, x):
-    return 1.0/(x*x*x)
-
-def inverse_exp(c, x):
-    return math.exp(-c*x)
-
-def gauss(c, x):
-    return math.exp(-c*x*x/2.0)
 
 class SvAttractorNode(bpy.types.Node, SverchCustomTreeNode):
     '''Attraction vectors calculator'''
@@ -67,10 +53,10 @@ class SvAttractorNode(bpy.types.Node, SverchCustomTreeNode):
             ("gauss", "Gauss - Exp(-R^2/2)", "", 4)
         ]
 
+    @throttled
     def update_type(self, context):
         self.inputs['Direction'].hide_safe = (self.attractor_type == 'Point')
         self.inputs['Coefficient'].hide_safe = (self.falloff_type not in ['inverse_exp', 'gauss'])
-        updateNode(self, context)
 
     attractor_type: EnumProperty(
         name="Attractor type", items=types, default='Point', update=update_type)
